Move Compose UI LazyLayout to codegen
